WIP getting hugo to like guides content
Hugo has weird behaviour. When ugly URLs are turned on, and a top-level
directory in `content` has a dot in it, it barfs with e.g. `Error: Error
building site: open public/2.2: is a directory` or similar. Moving
content into subfolders get around it but will require a whole load of
alias creation, let's see if that's worth it.